refactor(core): abstract RetrofitError.Kind.Network exceptions to own class
