+++ clouddriver-core/src/main/groovy/com/netflix/spinnaker/clouddriver/exceptions/SpinnakerHttpException.java
@@ -29,7 +29,7 @@ import retrofit.client.Response;
  */
 @Getter
 @NonnullByDefault
-public final class SpinnakerHttpException extends SpinnakerNetworkException {
+public final class SpinnakerHttpException extends SpinnakerServerException {
   private final Response response;
 
   public SpinnakerHttpException(RetrofitError e) {
